massabalans/requirements.txt
pandas
pandas_gbq
protobuf
python-dotenv
Requests
tabulate